Extract MetricsSelector
