+++ src/pages/dashboard/components/MetricsSelector/utils/metricsReducer.ts
