+++ src/store/reducers.ts
@@ -1,5 +1,5 @@
 import { reducer as weatherReducer } from '../Features/Weather/reducer';
-import { reducer as metricsReducer } from '../pages/dashboard/utils';
+import { reducer as metricsReducer } from '../pages/dashboard/components/MetricsSelector/utils';
 
 export default {
   weather: weatherReducer,
